Minor adjustment to wands of wishing
* They now always spawn with 1 charge.
* They now only gain 1 charge when using (noncursed) charging.
* However, they can be charged like any other wand, up to 7 times,
  but with a chance to explode when recharging more than 1 time.

What this means in practice is roughly the same amount of wishes,
but players now have to choose if they are willing to risk an explosion,
losing a wish in the process, or wrest a final guranteed wish after recharging
once.

Monsters will never recharge wands above the 1 recharge count, as previously.

Wished-for wands of wishing now have a recharge count of 7, for obvious reasons.